+++ src/utils/react/StringTableDisplay.tsx
